refactoring, make service autonomous

+++ raptor-http-api/src/main/java/org/createnet/raptor/http/api/ObjectApi.java
@@ -23,7 +23,10 @@ import java.net.URI;
 import java.util.List;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
+import javax.ws.rs.ForbiddenException;
 import javax.ws.rs.GET;
+import javax.ws.rs.InternalServerErrorException;
+import javax.ws.rs.NotFoundException;
 import javax.ws.rs.POST;
 import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
@@ -32,7 +35,10 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+import org.createnet.raptor.auth.authentication.Authentication;
+import org.createnet.raptor.auth.authorization.Authorization;
 import org.createnet.raptor.models.objects.ServiceObject;
+import org.createnet.raptor.search.Indexer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.createnet.raptor.search.query.impl.es.ObjectQuery;
@@ -59,7 +65,12 @@ public class ObjectApi extends AbstractApi {
             @QueryParam("limit") Integer limit,
             @QueryParam("offset") Integer offset
     ) {
-        return objectManager.list(offset, limit);
+
+        if (!auth.isAllowed(Authorization.Permission.List)) {
+            throw new ForbiddenException("Cannot list objects");
+        }
+
+        return objectManager.list(auth.getUser().getUserId(), offset, limit);
     }
 
     @POST
@@ -68,13 +79,39 @@ public class ObjectApi extends AbstractApi {
     @ApiResponses(value = {
         @ApiResponse(code = 200, message = "Ok")
         ,
-    @ApiResponse(code = 403, message = "Forbidden")
+        @ApiResponse(code = 403, message = "Forbidden")
         ,
-    @ApiResponse(code = 500, message = "Internal error")
+        @ApiResponse(code = 500, message = "Internal error")
     })
-    public Response create(ServiceObject obj) {
-        objectManager.create(obj);
-        return Response.created(URI.create("/" + obj.id)).entity(obj.toJSON()).build();
+    public ServiceObject create(ServiceObject obj) {
+
+        if (!auth.isAllowed(Authorization.Permission.Create)) {
+            throw new ForbiddenException("Cannot create object");
+        }
+
+        obj.userId = auth.getUser().getUserId();
+
+        try {
+            obj = objectManager.create(obj);
+        } catch (Indexer.IndexerException ex) {
+
+            logger.error("Indexing error occured", ex);
+            logger.warn("Removing object {} from storage", obj.id);
+
+            objectManager.delete(obj.id);
+
+            throw new InternalServerErrorException("Failed to index device");
+        }
+
+        //move to broker system events
+        boolean sync = syncObject(obj, Authentication.SyncOperation.CREATE);
+        if (!sync) {
+            logger.error("Auth sync failed, aborting creation of object {}", obj.id);
+            objectManager.delete(obj.id);
+            throw new InternalServerErrorException("Failed to sync device");
+        }
+
+        return obj;
     }
 
     @PUT
@@ -92,7 +129,33 @@ public class ObjectApi extends AbstractApi {
             @PathParam("id") String id,
             ServiceObject obj
     ) {
-        objectManager.update(id, obj);
+
+        ServiceObject storedObj = objectManager.load(id);
+
+        if (obj.id == null || obj.id.isEmpty()) {
+            obj.id = storedObj.id;
+        }
+
+        if (!storedObj.id.equals(obj.id)) {
+            throw new NotFoundException("Request id does not match payload defined id");
+        }
+
+        if (!auth.isAllowed(obj, Authorization.Permission.Update)) {
+            throw new ForbiddenException("Cannot update object");
+        }
+
+        if (!storedObj.userId.equals(auth.getUser().getUserId())) {
+            logger.warn("User {} tried to update object {} owned by {}", auth.getUser().getUserId(), storedObj.id, storedObj.userId);
+            throw new NotFoundException();
+        }
+
+        obj = objectManager.update(obj);
+
+        boolean sync = syncObject(obj, Authentication.SyncOperation.UPDATE);
+        if (!sync) {
+            throw new InternalServerErrorException("Failed to sync device");
+        }
+
         return obj;
     }
 
@@ -105,7 +168,14 @@ public class ObjectApi extends AbstractApi {
     @ApiResponse(code = 403, message = "Forbidden")
     })
     public ServiceObject load(@PathParam("id") String id) {
-        return objectManager.load(id);
+        
+        ServiceObject obj = objectManager.load(id);
+        
+        if (!auth.isAllowed(obj, Authorization.Permission.Read)) {
+            throw new ForbiddenException("Cannot read object");
+        }
+        
+        return obj;
     }
 
     @DELETE
@@ -119,7 +189,22 @@ public class ObjectApi extends AbstractApi {
     @ApiResponse(code = 500, message = "Internal error")
     })
     public Response delete(@PathParam("id") String id) {
+        
+        
+        ServiceObject obj = objectManager.load(id);
+
+        if (!auth.isAllowed(obj, Authorization.Permission.Delete)) {
+            throw new ForbiddenException("Cannot delete object");
+        }
+
         objectManager.delete(id);
+        
+        boolean sync = syncObject(obj, Authentication.SyncOperation.DELETE);
+        if (!sync) {
+            logger.error("Auth sync failed, aborting deletion of object {}", obj.id);
+            throw new InternalServerErrorException("Failed to sync device");
+        }
+        
         return Response.status(Response.Status.OK).build();
     }
 
@@ -134,7 +219,18 @@ public class ObjectApi extends AbstractApi {
     @ApiResponse(code = 403, message = "Forbidden")
     })
     public List<ServiceObject> search(ObjectQuery query) {
-        return objectManager.search(query);
+        
+        
+        if (!auth.isAllowed(Authorization.Permission.List)) {
+            throw new ForbiddenException("Cannot search for objects");
+        }
+        
+        //@TODO load accessible objects from auth service api!
+        query.setUserId(auth.getUser().getUserId());
+        
+        List<ServiceObject> list = objectManager.search(query);
+        
+        return list;
     }
 
 }
